Organized the tests.
